requirement.txt
streamlit==1.28.1
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.3
pillow==10.0.0
tensorflow==2.13.0
scikit-learn==1.3.0
joblib==1.3.1
requests==2.31.0